project/requirements.txt
# Core image processing
opencv-python>=4.8.0
scipy>=1.11.0
Pillow>=10.0.0
numpy>=1.24.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face — OCR + NLP models
transformers>=4.40.0
datasets>=2.18.0
evaluate>=0.4.1
accelerate>=0.27.0
huggingface_hub>=0.20.0

# OCR pipeline
kraken>=4.3.0
tqdm>=4.65.0
editdistance>=0.6.0

# Metrics
jiwer>=3.0.3

# Web app
streamlit>=1.30.0
python-docx>=1.0.0

# LLM correction (optional — only needed for Gemini stage)
google-genai>=1.0.0

# Misc
scikit-learn>=1.3.0